Disable compile buttons if no server connection
