+++ gui/src/app/CompilationServerConnectionControl/CompilationServerConnectionControl.tsx
@@ -3,13 +3,7 @@ import { Cancel, Check } from "@mui/icons-material";
 import CloseableDialog, {
   useDialogControls,
 } from "@SpComponents/CloseableDialog";
-import {
-  FunctionComponent,
-  useCallback,
-  useContext,
-  useEffect,
-  useState,
-} from "react";
+import { FunctionComponent, useCallback, useContext } from "react";
 import ConfigureCompilationServerDialog from "./ConfigureCompilationServerDialog";
 import IconButton from "@mui/material/IconButton";
 import Typography from "@mui/material/Typography";
@@ -28,8 +22,8 @@ type CompilationServerConnectionControlProps = {
 const CompilationServerConnectionControl: FunctionComponent<
   CompilationServerConnectionControlProps
 > = () => {
-  const { stanWasmServerUrl } = useContext(CompileContext);
-  const { isConnected, retryConnection } = useIsConnected(stanWasmServerUrl);
+  const { stanWasmServerUrl, isConnected, retryConnection } =
+    useContext(CompileContext);
 
   const {
     handleOpen: openDialog,
@@ -78,34 +72,6 @@ export const serverTypeForUrl = (url: string): ServerType => {
     : url === localCompilationServerUrl
       ? "local"
       : "custom";
-};
-
-const useIsConnected = (stanWasmServerUrl: string) => {
-  const probeUrl = `${stanWasmServerUrl}/probe`;
-  const [isConnected, setIsConnected] = useState<boolean>(false);
-  const [retryCode, setRetryCode] = useState<number>(0);
-  const retryConnection = useCallback(() => {
-    setRetryCode((r) => r + 1);
-  }, []);
-  useEffect(() => {
-    setIsConnected(false);
-    if (!probeUrl.startsWith("http://") && !probeUrl.startsWith("https://")) {
-      // important to do this check because otherwise fetch may succeed because
-      // the server of this web app may respond with success
-      return;
-    }
-    (async () => {
-      try {
-        const response = await fetch(probeUrl);
-        if (response.status === 200) {
-          setIsConnected(true);
-        }
-      } catch (err) {
-        setIsConnected(false);
-      }
-    })();
-  }, [probeUrl, retryCode]);
-  return { isConnected, retryConnection };
 };
 
 export default CompilationServerConnectionControl;
