+++ gui/src/app/CompileContext/CompileContextProvider.tsx
@@ -15,6 +15,34 @@ type CompileContextProviderProps = {
   // none
 };
 
+const useIsConnected = (stanWasmServerUrl: string) => {
+  const probeUrl = `${stanWasmServerUrl}/probe`;
+  const [isConnected, setIsConnected] = useState<boolean>(false);
+  const [retryCode, setRetryCode] = useState<number>(0);
+  const retryConnection = useCallback(() => {
+    setRetryCode((r) => r + 1);
+  }, []);
+  useEffect(() => {
+    setIsConnected(false);
+    if (!probeUrl.startsWith("http://") && !probeUrl.startsWith("https://")) {
+      // important to do this check because otherwise fetch may succeed because
+      // the server of this web app may respond with success
+      return;
+    }
+    (async () => {
+      try {
+        const response = await fetch(probeUrl);
+        if (response.status === 200) {
+          setIsConnected(true);
+        }
+      } catch (err) {
+        setIsConnected(false);
+      }
+    })();
+  }, [probeUrl, retryCode]);
+  return { isConnected, retryConnection };
+};
+
 const initialStanWasmServerUrl =
   localStorage.getItem("stanWasmServerUrl") || publicCompilationServerUrl;
 
@@ -83,6 +111,8 @@ export const CompileContextProvider: FunctionComponent<
     stanWasmServerUrl,
   ]);
 
+  const { isConnected, retryConnection } = useIsConnected(stanWasmServerUrl);
+
   return (
     <CompileContext.Provider
       value={{
@@ -94,6 +124,8 @@ export const CompileContextProvider: FunctionComponent<
         setValidSyntax,
         stanWasmServerUrl,
         setStanWasmServerUrl,
+        isConnected,
+        retryConnection,
       }}
     >
       {children}
